edge-kit/edge-kit/api/requirements.txt
Flask>=2.0.0
psutil>=5.9.0
PyYAML>=6.0